sum to god file
