+++ tests/test_agent_core_p2_reasoner.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from types import SimpleNamespace
 from typing import Any, cast
 
-from agent.core.reasoner import DefaultReasoner
+from agent.core.passive_turn import DefaultReasoner
 from agent.core.runtime_support import LLMServices, ToolDiscoveryState
 from agent.looping.ports import LLMConfig
 from agent.provider import LLMResponse, ToolCall
@@ -251,7 +251,7 @@ def test_default_reasoner_unlocks_tool_search_visibility():
 
 def test_default_reasoner_preflight_includes_deferred_tool_names():
     """调用方（如 _run_agent_loop）负责注入 deferred tools hint；run() 本身不再自动注入。"""
-    from agent.core.reasoner import build_turn_injection_prompt
+    from agent.core.passive_turn import build_turn_injection_prompt
     from agent.prompting import build_context_frame_content, build_context_frame_message
     from agent.prompting import PromptSectionRender
